requirements.txt
numpy~=1.21.4
tensorflow~=2.7.0
tqdm~=4.62.3
streamlit~=1.2.0
pandas~=1.3.4
Pillow~=8.4.0
sklearn~=0.0
scikit-learn~=1.0.1
opencv-python~=4.5.4.60